requirements.txt
aiocontextvars==0.2.0
aiofiles==0.4.0
aiohttp==3.4.4
aioredis==1.2.0
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.17.0
atomicwrites==1.2.1
attrs==18.2.0
chardet==3.0.4
gino==0.8.0
gunicorn==19.9.0
hiredis==0.3.0
httptools==0.0.11
idna==2.7
more-itertools==4.3.0
multidict==4.5.0
pluggy==0.8.0
py==1.7.0
PyJWT==1.6.4
pytest==4.0.1
pytest-sanic==0.1.15
python-dotenv==0.9.1
sanic==0.8.3
sanic-jwt==1.2.0
sanic-openapi==0.4.0
six==1.11.0
SQLAlchemy==1.2.12
ujson==1.35
uvloop==0.11.3
websockets==5.0.1
yarl==1.2.6
